+++ admin/contact_us.php
@@ -7,6 +7,7 @@ if (isset($_GET['type']) && $_GET['type'] != '') {
 		$id = get_safe_value($con, $_GET['id']);
 		$delete_sql = "delete from contact_us where id='$id'";
 		mysqli_query($con, $delete_sql);
+		set_flash('msg', 'Contact query deleted successfully ✅');
 	}
 }
 
@@ -48,7 +49,7 @@ $res = mysqli_query($con, $sql);
 											<td><?php echo $row['added_on'] ?></td>
 											<td>
 												<?php
-												echo "<span class='badge badge-delete'><a href='?type=delete&id=" . $row['id'] . "'>Delete</a></span>";
+												echo "<span class='badge badge-delete'><a href='?type=delete&id=" . $row['id'] . "'onclick=\"return confirm('Are you sure you want to delete this query?')\">Delete</a></span>";
 												?>
 											</td>
 										</tr>
@@ -56,6 +57,11 @@ $res = mysqli_query($con, $sql);
 								</tbody>
 							</table>
 						</div>
+						<?php if ($msg = get_flash('msg')) { ?>
+							<div id="myToast" class="myToast__msg">
+								<?php echo $msg; ?>
+							</div>
+						<?php } ?>
 					</div>
 				</div>
 			</div>
